perf: split page into two components
